+++ src/main/java/io/numaproj/numaflow/reducestreamer/ActorResponseType.java
@@ -1,6 +1,6 @@
 package io.numaproj.numaflow.reducestreamer;
 
-public enum ActorResponseType {
+enum ActorResponseType {
     // EOF_RESPONSE indicates that the actor response contains an EOF reduce response without real data.
     EOF_RESPONSE,
     // READY_FOR_CLEAN_UP_SIGNAL indicates that the actor has finished sending responses and now ready to be cleaned up.
