+++ src/main/java/io/numaproj/numaflow/reducestreamer/Service.java
@@ -9,11 +9,9 @@ import io.grpc.stub.StreamObserver;
 import io.numaproj.numaflow.reduce.v1.ReduceGrpc;
 import io.numaproj.numaflow.reduce.v1.ReduceOuterClass;
 import io.numaproj.numaflow.reducestreamer.model.IntervalWindow;
-import io.numaproj.numaflow.reducestreamer.model.IntervalWindowImpl;
 import io.numaproj.numaflow.reducestreamer.model.Metadata;
-import io.numaproj.numaflow.reducestreamer.model.MetadataImpl;
-import io.numaproj.numaflow.reducestreamer.user.ReduceStreamer;
-import io.numaproj.numaflow.reducestreamer.user.ReduceStreamerFactory;
+import io.numaproj.numaflow.reducestreamer.model.ReduceStreamer;
+import io.numaproj.numaflow.reducestreamer.model.ReduceStreamerFactory;
 import io.numaproj.numaflow.shared.GrpcServerUtils;
 import lombok.extern.slf4j.Slf4j;
 
@@ -24,7 +22,6 @@ import static io.numaproj.numaflow.reduce.v1.ReduceGrpc.getReduceFnMethod;
 
 @Slf4j
 class Service extends ReduceGrpc.ReduceImplBase {
-
     public static final ActorSystem reduceActorSystem = ActorSystem.create("reducestream");
 
     private ReduceStreamerFactory<? extends ReduceStreamer> reduceStreamerFactory;
