+++ src/main/java/io/numaproj/numaflow/reducestreamer/model/Datum.java
@@ -5,7 +5,6 @@ import java.time.Instant;
 /**
  * Datum contains methods to get the payload information.
  */
-
 public interface Datum {
     /**
      * method to get the payload value
