+++ src/main/java/io/numaproj/numaflow/reducestreamer/model/ReduceStreamerFactory.java
@@ -1,10 +1,8 @@
-package io.numaproj.numaflow.reducestreamer.user;
-
+package io.numaproj.numaflow.reducestreamer.model;
 
 /**
  * ReduceStreamerFactory is the factory for ReduceStreamer.
  */
-
 public abstract class ReduceStreamerFactory<ReduceStreamerT extends ReduceStreamer> {
     public abstract ReduceStreamerT createReduceStreamer();
 }
