+++ src/test/java/io/numaproj/numaflow/reducestreamer/ServerTest.java
@@ -17,9 +17,9 @@ import io.numaproj.numaflow.reduce.v1.ReduceGrpc;
 import io.numaproj.numaflow.reduce.v1.ReduceOuterClass;
 import io.numaproj.numaflow.reducestreamer.model.Datum;
 import io.numaproj.numaflow.reducestreamer.model.Message;
-import io.numaproj.numaflow.reducestreamer.user.OutputStreamObserver;
-import io.numaproj.numaflow.reducestreamer.user.ReduceStreamer;
-import io.numaproj.numaflow.reducestreamer.user.ReduceStreamerFactory;
+import io.numaproj.numaflow.reducestreamer.model.OutputStreamObserver;
+import io.numaproj.numaflow.reducestreamer.model.ReduceStreamer;
+import io.numaproj.numaflow.reducestreamer.model.ReduceStreamerFactory;
 import io.numaproj.numaflow.shared.GrpcServerUtils;
 import org.junit.After;
 import org.junit.Before;
